requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scipy>=1.9

# Deep Learning
torch>=2.0
torchvision>=0.15

# Image processing
Pillow>=9.4

# Evaluation / metrics
scikit-learn>=1.2

# Utilities
tqdm>=4.65
pyyaml>=6.0
matplotlib>=3.7
